Compute individual coalescence counts at each node (MapMutations)
During MapMutations we have all the information we need to track
how many individuals are coalescing (for diploids) at each node
that is visited/created. We pay essentially no cost for computing
this, and its about 1% increase in GRG size.

This is now accessible from grgl::NodeData via the
numIndividualCoals field, and is extremely useful for any
computations that requires {0, 1, 2} form for the matrix version
of the computation.

Added the copying of NodeData in GRG merge operation, since we
now have more than populationId (it should have always been copied)