Change grg split: -s for size-based splitting, -f for ranges
You must specify either "-s" or "-f". The former takes an number
as argument, which specifies how big each GRG should be when
splitting (in BP, or cM if --rec-map is specified). The latter
takes a filename, which is a text file with two columns like:
  start end
  0 10000
  20000 21000
  ...
which specifies ranges to create GRGs for.