matmul() support for smaller datatypes, including bitvectors
Using a smaller numpy datatype saves memory and also speeds up
computation. For general "counting" computation, you cannot use
anything smaller than numpy.int32 because a GRG can have up to
(2**32)-2 nodes in the graph. However, in special cases the int16
and int8 could be useful.

One special case that is useful is the boolean problem, where the
input as a vector with a single 1 and the rest are zeros. This can
be used to explicitly retrieve thegenotype matrix, for example.
In this case, we only need boolean values, and matmul now stores
all node values packed into a bitvectors.