Node initialization option for `matmul()` (#65)
* Added matmul(init=...) feature

Used to initialize node values prior to running the matrix
multiplication. Can be used to compute the X^t * X product needed
for GWAS, by specifying "init='xtx'".

For general initialization, you can either pass in a vector (one
value per row of the input matrix, which will get copied to all
nodes for that row) or a matrix (which must be KxT, where K is the
number of rows in the input matrix and T is the total number of
nodes in the GRG).