Faster GRG merge, retains more hierarchy
Much faster GRG merge, with the expectation for lower RAM usage
on large GRGs as well. Previously we mapped based on the samples
beneath a node, which could change the subtree beneath a node,
sometimes removing hierarchy. It is better to keep hierarchy
at the cost of a little bit larger graph. Also, the new algorithm
does not propoagate data transitively in the graph, it only looks
at immediate children, which makes it much more efficient.