Optionally split files by variant count
By default, IGD and BGEN inputs will be split by variant count
now. This means each segment that we build a graph for has roughly
the same number of variants, which should improve the parallel
behavior and might improve the file size.

VCF is very inefficient, currently, and splitting by variant will
probably only make it worse, so we leave it as base-pair by
default.

Users can specify this behavior, when building segments or for a
specific range, by using the "b" (base pair) or "v" (variant)
suffix on a range. E.g., "0.5-0.51v" will use the variants in the
range 50% to 51%, of all variants.