Split genome by variant ranges (#71)
* Optionally split files by variant count

By default, IGD and BGEN inputs will be split by variant count
now. This means each segment that we build a graph for has roughly
the same number of variants, which should improve the parallel
behavior and might improve the file size.

VCF is very inefficient, currently, and splitting by variant will
probably only make it worse, so we leave it as base-pair by
default.

Users can specify this behavior, when building segments or for a
specific range, by using the "b" (base pair) or "v" (variant)
suffix on a range. E.g., "0.5-0.51v" will use the variants in the
range 50% to 51%, of all variants.

* Make range denomormalization more type safe

Fix bugs related to how range coverage is computed, and the
denormalization from a FloatRange to a (new type) IntRange.

* IGD needs to stop at position boundaries

When using variant ranges, don't overlap positions between ranges.
I.e., if the previous range was (A, B) and the new one is (B, C),
then if the variants with indexes (B-1) and B have the same
position we want them to be emitted in the first range, not the
second (or we'll get duplicate Mutations)