Proper missing data support
The previous missing data support was pretty weak, it just generated
additional Mutation nodes that had alleles that indicated they were
missing data. This made it pretty tricky to do useful calculations.
Also, those Mutations took up a bunch of space that wasn't necessary.

Instead, we have the new concept of a "Missingness node". So there
are four kinds of nodes:
1. Sample nodes, numbered 0-(N-1), where N is the numer of haploid
   samples.
2. Mutation nodes, which can be any node (including Sample nodes)
   and just mean there is a Mutation attached to the node.
3. Empty nodes, "nothing special" or attached to the node.
4. Missingness nodes, which can be any node similar to a Mutation
   node, but they have no meta-data associated with them. They are
   however associated with a single Mutation node. So a multi-
   allelic site might have 3 Mutations (4 alleles total, including
   the REF) and all 3 of those Mutations will have distinct
   MutationIds, and each will be associated to a single missingness
   node that represents the missingness for that site.

Still a work-in-progress.