Bug fix for rare scenario with IGD variant ranges
Variant-based ranges are used internally when IGD is converted to
GRG, because it is pretty efficient. The downside is that the
variant ranges can overlap a position, if that position is multi-
allelic. I had code to handle this, but it didn't properly
consider the end position case w.r.t. countMutations(), meaning
you could ask "countMutations()" and get a number smaller or
larger than you would actually traverse.