Graph transformation to add hierarchy and compress data
The current GRG construction algorithm is extremely fast, and
generates graphs that are fairly "wide" (large branch factor, the
height of the graph is usually small, ~20 or so). We know that
adding hierarchy is valuable for compressing. This change adds
a simple transformation that looks at siblings in the graph and
if they have a lot of shared children it will introduce hierarchy
to capture those shared children.