Maintain mutations on ancestral path where possible
The "limited multi-tree" we use during construction is convenient
for reasoning about GRG construction, but in some cases it can
create extra mutation nodes "off to the side" when we could have,
in fact, created one in the current tree hierarchy. This detects
that scenario and makes sure we always include it in the tree
itself when possible.