Fix bug where valid merge is rejected; add lots of tests
With multi-allelic data, we can have two GRGs that have an allele
at the same position, but are still valid to merge. This happens
when that position starts one GRG and ends the other. This was
rejected as an invalid GRG merge previously.

I've added an end-to-end test for this, and also a bunch of unit
tests for the (now generalized) position/range checking functions.