Fix parallelism load balancing problem
Pool.map (and imap_unordered, that we use) does not in fact use a
default chunksize=1 as per the documentation, but attempts to
calculate a "good" value for you by default. This causes serious
load imbalances for large GRG constructions, if two regions of
the genome that are high in diversity end up in the same map()
batch.